refactor(ort-run): Extract the effective PURL query building
Run issue filtering and sorting also needs the effective PURL in SQL.
Extract the PURL building functions to an own module so the issues query 
can use the same analyzer and curated PURL precedence as both 
vulnerability queries instead of adding a third copy of the functions.
