+++ services/ort-run/src/main/kotlin/OrtRunVulnerabilityQueryBuilder.kt
@@ -27,10 +27,6 @@ import org.eclipse.apoapsis.ortserver.dao.repositories.advisorrun.AdvisorRunsTab
 import org.eclipse.apoapsis.ortserver.dao.repositories.advisorrun.ResolvedVulnerabilitiesTable
 import org.eclipse.apoapsis.ortserver.dao.repositories.advisorrun.VulnerabilitiesTable
 import org.eclipse.apoapsis.ortserver.dao.repositories.advisorrun.VulnerabilityReferencesTable
-import org.eclipse.apoapsis.ortserver.dao.repositories.analyzerjob.AnalyzerJobsTable
-import org.eclipse.apoapsis.ortserver.dao.repositories.analyzerrun.AnalyzerRunsTable
-import org.eclipse.apoapsis.ortserver.dao.repositories.analyzerrun.PackagesAnalyzerRunsTable
-import org.eclipse.apoapsis.ortserver.dao.repositories.analyzerrun.PackagesTable
 import org.eclipse.apoapsis.ortserver.dao.repositories.ortrun.OrtRunsTable
 import org.eclipse.apoapsis.ortserver.dao.tables.shared.IdentifiersTable
 import org.eclipse.apoapsis.ortserver.dao.utils.applyILike
@@ -38,14 +34,12 @@ import org.eclipse.apoapsis.ortserver.model.VulnerabilityFilters
 import org.eclipse.apoapsis.ortserver.model.util.ComparisonOperator
 import org.eclipse.apoapsis.ortserver.model.util.FilterOperatorAndValue
 
-import org.jetbrains.exposed.v1.core.CustomFunction
 import org.jetbrains.exposed.v1.core.Expression
 import org.jetbrains.exposed.v1.core.ExpressionWithColumnType
 import org.jetbrains.exposed.v1.core.ExpressionWithColumnTypeAlias
 import org.jetbrains.exposed.v1.core.JoinType
 import org.jetbrains.exposed.v1.core.Op
 import org.jetbrains.exposed.v1.core.QueryAlias
-import org.jetbrains.exposed.v1.core.TextColumnType
 import org.jetbrains.exposed.v1.core.alias
 import org.jetbrains.exposed.v1.core.and
 import org.jetbrains.exposed.v1.core.concat
@@ -53,7 +47,6 @@ import org.jetbrains.exposed.v1.core.dao.id.EntityID
 import org.jetbrains.exposed.v1.core.eq
 import org.jetbrains.exposed.v1.core.inList
 import org.jetbrains.exposed.v1.core.innerJoin
-import org.jetbrains.exposed.v1.core.min
 import org.jetbrains.exposed.v1.core.not
 import org.jetbrains.exposed.v1.core.notExists
 import org.jetbrains.exposed.v1.core.notInList
@@ -69,7 +62,19 @@ internal fun buildOrtRunVulnerabilityQueryContext(
 
     val bp = buildOrtRunBasePairs(ortRunId, filters.advisors)
     val ratings = buildOrtRunRatingsByVulnerability(bp, ratingAlias)
-    val purls = buildOrtRunPurlByRunIdentifier(bp)
+
+    val riOrtRunId = bp.alias[bp.ortRunId].alias("ri_ort_run_id")
+    val riIdentifierId = bp.alias[bp.identifierId].alias("ri_identifier_id")
+    val runIdPairs = bp.alias
+        .select(riOrtRunId, riIdentifierId)
+        .withDistinct()
+        .alias("run_identifier_pairs")
+
+    val purls = buildPurlByRunIdentifier(
+        runIdPairs,
+        runIdPairs[riOrtRunId],
+        runIdPairs[riIdentifierId]
+    )
 
     return buildOrtRunFinalQuery(bp, ratings, purls, ratingAlias, filters)
 }
@@ -153,60 +158,6 @@ private fun buildOrtRunRatingsByVulnerability(
         .select(VulnerabilityReferencesTable.vulnerabilityId, ratingAlias)
         .groupBy(VulnerabilityReferencesTable.vulnerabilityId)
         .alias("run_ratings_by_vulnerability")
-}
-
-private fun buildOrtRunPurlByRunIdentifier(bp: OrtRunBasePairsResult): PurlResult {
-    val riOrtRunId = bp.alias[bp.ortRunId].alias("run_ri_ort_run_id")
-    val riIdentifierId = bp.alias[bp.identifierId].alias("run_ri_identifier_id")
-    val runIdPairs = bp.alias
-        .select(riOrtRunId, riIdentifierId)
-        .withDistinct()
-        .alias("run_identifier_pairs")
-
-    val riOrtRunIdCol = runIdPairs[riOrtRunId]
-    val riIdentifierIdCol = runIdPairs[riIdentifierId]
-
-    val bpOrtRunId = riOrtRunIdCol.alias("run_bp_ort_run_id")
-    val bpIdentifierId = riIdentifierIdCol.alias("run_bp_identifier_id")
-    val basePurlMin = PackagesTable.purl.min().alias("run_base_purl")
-
-    val basePurl = AnalyzerJobsTable
-        .innerJoin(AnalyzerRunsTable)
-        .innerJoin(PackagesAnalyzerRunsTable)
-        .innerJoin(PackagesTable)
-        .join(runIdPairs, JoinType.INNER) {
-            (AnalyzerJobsTable.ortRunId eq riOrtRunIdCol) and
-                (PackagesTable.identifierId eq riIdentifierIdCol)
-        }
-        .select(bpOrtRunId, bpIdentifierId, basePurlMin)
-        .groupBy(bpOrtRunId, bpIdentifierId)
-        .alias("run_base_purl_by_run_identifier")
-
-    val curated = buildCuratedPurlByRunIdentifier(runIdPairs, riOrtRunIdCol, riIdentifierIdCol)
-
-    val purl = CustomFunction(
-        "COALESCE",
-        TextColumnType(),
-        curated.alias[curated.purl],
-        basePurl[basePurlMin],
-        stringLiteral("")
-    ).alias("run_purl")
-    val outOrtRunId = riOrtRunIdCol.alias("run_purl_ort_run_id")
-    val outIdentifierId = riIdentifierIdCol.alias("run_purl_identifier_id")
-
-    val query = runIdPairs
-        .join(basePurl, JoinType.LEFT) {
-            (riOrtRunIdCol eq basePurl[bpOrtRunId]) and
-                (riIdentifierIdCol eq basePurl[bpIdentifierId])
-        }
-        .join(curated.alias, JoinType.LEFT) {
-            (riOrtRunIdCol eq curated.alias[curated.ortRunId]) and
-                (riIdentifierIdCol eq curated.alias[curated.identifierId])
-        }
-        .select(outOrtRunId, outIdentifierId, purl)
-        .alias("run_purl_by_run_identifier")
-
-    return PurlResult(query, outOrtRunId, outIdentifierId, purl)
 }
 
 private fun buildOrtRunFinalQuery(
